+++ implementation/js/src/reusable/editors/index.js
@@ -10,9 +10,9 @@
  * 		field: ...
  * }
  * type (see predefined/parts/editors/README.md)
- * label
- * help
- * tooltip
+ * label (not sanitized)
+ * help (not sanitized)
+ * tooltip (not sanitized)
  * placeholder
  * value: default value
  *
@@ -23,6 +23,7 @@
  * 	labelField
  * 	valueField
  * 	remote: app.remote() options for fetching the options.data
+ * 	extract: 'key.sub.path.' or fn to get data array from remote
  * }
  *
  * //single checkbox only
@@ -68,6 +69,7 @@
  * @updated 2014.02.26 [Bootstrap 3.1+]
  * @updated 2015.12.07 [awesome-bootstrap-checkbox & radio]
  * @updated 2016.11.16
+ * @updated 2017.04.27
  * @version 1.2.1
  */
 
@@ -130,13 +132,15 @@
 						labelField: 'label'
 					}, options.options);
 
-					var choices = options.options; //for easy reference within extractChoices()
-					var extractChoices = function (data){
+					var choices = options.options; //for easy reference within getKVChoices()
+					var getKVChoices = function (data){
 						if(_.isObject(data[0])){
+							//a. pick kv from each object in array
 							data = _.map(data, function(c){
 								return {value: c[choices.valueField], label: c[choices.labelField]};
 							});
 						}else {
+							//b. generate k from each v in array
 							data = _.map(data, function(c){
 								return {value: c, label: _.string.titleize(c)};
 							});
@@ -145,7 +149,17 @@
 					};
 
 					var prepareChoices = function (choices){
+						//extract real choices from data obj (non-array)
+						if(choices.extract){
+							if(!_.isArray(choices.data)){
+								if(_.isString(choices.extract))
+									choices.data = app.extract(choices.extract, choices.data);
+								else if(_.isFunction(choices.extract))
+									choices.data = choices.extract(choices.data);							
+							}							
+						}
 
+						//see if (extracted) data is grouped (auto-detect)
 						if(!_.isArray(choices.data)){
 							choices.grouped = true;
 						}
@@ -153,11 +167,11 @@
 						if(choices.grouped){
 							//select (grouped)
 							_.each(choices.data, function(array, group){
-								choices.data[group] = extractChoices(array);
+								choices.data[group] = getKVChoices(array);
 							});
 						}else {
 							//select, radios, checkboxes
-							choices.data = extractChoices(choices.data);
+							choices.data = getKVChoices(choices.data);
 						}
 
 						return choices;
@@ -169,7 +183,6 @@
 						this.listenToOnce(this, 'render', function(){
 							var that = this;
 							app.remote(choices.remote).done(function(data){
-
 								//Warning: to leave less config overhead, developers have no way to pre-process the choice data returned atm.
 								that.setChoices(data);
 							});
@@ -946,7 +959,7 @@
 						.removeClass(this.$el.data('type-class'))
 						.addClass(className)
 						.data('type-class', className);
-					this.ui.msg.html(msg.i18n());
+					this.ui.msg.html(msg.i18n()); //not sanitized
 
 				}else {
 					//clear
@@ -1000,9 +1013,9 @@
 
 	app.Util.Tpl.build('editor-basic-tpl', [
 		'{{#if label}}',
-			'<label class="control-label {{#if layout}}{{layout.label}}{{/if}}" for="{{uiId}}">{{i18n label}}</label>',
+			'<label class="control-label {{#if layout}}{{layout.label}}{{/if}}" for="{{uiId}}">{{{i18n label}}}</label>',
 		'{{/if}}',
-		'<div class="{{#if layout}}{{layout.field}}{{/if}}" data-toggle="tooltip" title="{{i18n tooltip}}">', //for positioning with the label.
+		'<div class="{{#if layout}}{{layout.field}}{{/if}}" data-toggle="tooltip" title="{{{i18n tooltip}}}">', //for positioning with the label.
 
 			//1. select
 			'{{#is type "select"}}',
@@ -1011,13 +1024,13 @@
 						'{{#each options.data}}',
 						'<optgroup label="{{i18n @key}}">',
 							'{{#each this}}',
-							'<option value="{{value}}">{{i18n label}}</option>',
+							'<option value="{{value}}">{{{i18n label}}}</option>',
 							'{{/each}}',
 						'</optgroup>',
 						'{{/each}}',
 					'{{else}}',
 						'{{#each options.data}}',
-						'<option value="{{value}}">{{i18n label}}</option>',
+						'<option value="{{value}}">{{{i18n label}}}</option>',
 						'{{/each}}',
 					'{{/if}}',
 				'</select>',
@@ -1034,7 +1047,7 @@
 							'<div class="{{../type}} {{#if ../options.inline}}{{../type}}-inline{{/if}}">',
 								//note that the {{if}} within a {{each}} will no longer impose +1 level down in the content scope. (after Handlebars v4)
 								'<input id="{{../uiId}}-{{@index}}" ui="input" name="{{#if ../fieldname}}{{../fieldname}}{{else}}{{../name}}{{/if}}{{#is ../type "checkbox"}}[]{{/is}}" type="{{../type}}" value={{value}}> ',
-								'<label for="{{../uiId}}-{{@index}}">{{i18n label}}</label>',
+								'<label for="{{../uiId}}-{{@index}}">{{{i18n label}}}</label>',
 							'</div>',
 						'{{/each}}',
 						'</div>',
@@ -1044,7 +1057,7 @@
 						'{{#is type "checkbox"}}',
 							//single checkbox
 							'<input id="{{uiId}}" ui="input" name="{{#if fieldname}}{{fieldname}}{{else}}{{name}}{{/if}}" type="checkbox" value="{{value}}"> ',
-							'<label for="{{uiId}}">{{i18n boxLabel}}</label>',
+							'<label for="{{uiId}}">{{{i18n boxLabel}}}</label>',
 						'{{else}}',
 							//normal field
 							'{{#is type "ro"}}',//read-only
@@ -1195,8 +1208,8 @@
 			'{{/is}}',
 
 			//msg & help
-			'{{#if help}}<span class="help-block editor-help-text" style="margin-bottom:0"><small>{{i18n help}}</small></span>{{/if}}',
-			'<span class="help-block editor-status-text input-error" ui="msg">{{i18n msg}}</span>',
+			'{{#if help}}<span class="help-block editor-help-text" style="margin-bottom:0"><small>{{{i18n help}}}</small></span>{{/if}}',
+			'<span class="help-block editor-status-text input-error" ui="msg">{{{i18n msg}}}</span>',
 		'</div>'
 	]);